requirementsnew.txt
pythran
cython
pybind11
asn1crypto
attrs
Automat
beautifulsoup4
cffi
constantly
cryptography
cssselect
enum34
hyperlink
idna
incremental
ipaddress
lxml
parsel
pyasn1
pyasn1-modules
pycparser
PyDispatcher
pymongo
pyOpenSSL
queuelib
Scrapy
service-identity
six
Twisted
w3lib
zope.interface
selenium
pandas
numpy
html2text
nltk
swifter
Sastrawi
openpyxl
fasttext
gensim
sklearn
nltk
keybert
mysql-connector-python